button javascript bug fix
