+++ app/javascript/controllers/index.js
@@ -7,8 +7,6 @@ import { application } from "controllers/application";
 import { eagerLoadControllersFrom } from "@hotwired/stimulus-loading"
 eagerLoadControllersFrom("controllers", application)
 
-import ActionButtonsController from "./action_buttons_controller";
-Stimulus.register("disable-button", ActionButtonsController)
 
 // Lazy load controllers as they appear in the DOM (remember not to preload controllers in import map!)
 // import { lazyLoadControllersFrom } from "@hotwired/stimulus-loading"
